CountdownTimerBackend/requirements.txt
Flask
Flask-SQLAlchemy
# Flask-SocketIO
flask-sockets
gevent
gevent-websocket
gunicorn
psycopg2-binary
python-dotenv
PyJWT
bcrypt
requests
